Network test: Don't launch Afina when doing integration tests
It's hard to determine when Afina is ready to handle incoming
connections. Therefore, leave the launching to the user and
provide options to specify the address and the port to connect to.

NOTE: there's currently no "delete" command in the parser, so the
test can't undo the changes it did. Passing tests more than once
is impossible for now.